+++ model/kinderen/kind.class.php
@@ -50,6 +50,9 @@ class Kind extends Record{
         return $query->fetch(PDO::FETCH_OBJ);
     }
     protected function delete(){
+    	if(count($this->getUitstapDeelnames()) > 0 || count($this->getKindVoogden()) > 0){
+    		return false;
+    	}
         $query = Database::getPDO()->prepare("DELETE FROM Kind WHERE Id = :id");
         $query->bindParam(':id', $this->Id, PDO::PARAM_INT);
         return $query->execute();
