+++ model/uitstappen/uitstap.class.php
@@ -48,6 +48,9 @@ class Uitstap extends Record{
         return $query->fetch(PDO::FETCH_OBJ);
      }
      protected function delete(){
+     	if($this->getAantalDeelnemers() > 0){
+     		return false;
+     	}
         $query = Database::getPDO()->prepare("DELETE FROM Uitstap WHERE Id = :id");
         $query->bindParam(':id', $this->Id, PDO::PARAM_INT);
         return $query->execute();
