+++ model/voogden/voogd.class.php
@@ -45,6 +45,9 @@ class Voogd extends Record{
         return $query->fetch(PDO::FETCH_OBJ);
      }
      protected function delete(){
+     	if(count($this->getKinderen()) > 0){
+     		return false;
+     	}
         $query = Database::getPDO()->prepare("DELETE FROM Voogd WHERE Id = :id");
         $query->bindParam(':id', $this->Id, PDO::PARAM_INT);
         return $query->execute();
